Move dependencies into a single file deps.ts
Specify the version tag for those which offer it.
